+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ## [0.63.0](https://github.com/returntocorp/semgrep/releases/tag/v0.63.0) - 08-25-2021
 
 ### Added
+- C#: support ellipsis in declarations (#3720)
 
 ### Fixed
 - Hack: improved support for metavariables (#3716)
@@ -23,7 +24,6 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Add help text to sarif formatter output if defined in metadata field.
 - Update shortDescription in sarif formatter output if defined in metadata field.
 - Add tags as defined in metadata field in addition to the existing tags.
-- C#: support ellipsis in declarations (#3720)
 
 ### Fixed
 - core: Fix parsing of numeric literals in rule files